Adding NCM support to Android USB gadget.
This change adds support for NCM connections into the Android USB gadget.
To enable such connections, you should run the following commands on
the Android device (as root):

  echo 0       > /sys/class/android_usb/android0/enable
  echo 12:34:45:12:34:45 > /sys/class/android_usb/android0/f_ncm/ethaddr_ncm
  echo adb,ncm > /sys/class/android_usb/android0/functions
  echo 1       > /sys/class/android_usb/android0/enable

Because these operations will cause the adb connection to be dropped, it's
recommended to do this programmatically (a shell script works fine).

This change has required the __init macros to be removed from many
parts of the NCM module. Previously, the NCM initialisation code was only
used when the device booted, and thus the kernel freed its memory
after boot. However, now that NCM functions can be bound and unbound
dynamically it's necessary to keep the NCM code around just in case.