Fix for NCM on Android 4.0.4.
NCM had worked on earlier kernel versions but no longer worked on Android 4.0.4.
The symptoms were that the USB device descriptor used string ID 3 for the MAC
address, which contained the word "Full" - obviously not a valid MAC address.

This turned out to be because the NCM device was using "usb_string_id()" to allocate
a string ID; this returned 3. Previously, android_bind() had called
the same function which had also returned 3, for the device name string.

This was because the Android driver was resetting the counter to 0 each time
a device was enabled. This is invalid behaviour, since some strings were
already allocated, and caused the conflicts we were seeing.

This is now fixed, but the fix may not be production-ready. Please see
the comment added by this commit for an explanation. A more sophisticated
fix may be required.